refactor: types (#22)
